requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
numpy==1.23.1
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.10
whitenoise==6.2.0
torch==1.12.0
torchvision==0.13.0
